fix: testing timeout usage (#401)
* chore: remove duplicate timeout definitions

The make test command was using the KNUU_TIMEOUT and -timeout variables.
I removed the -timeout flags as a default of 10m is sufficient.
We have a .env file that we can use for the KNUU_TIMEOUT so I removed it
from the command itself.

* fix: clean up ci test file

The merge_group trigger wasn't set so test won't be run on the merge
queue.
We were setting the KNUU_TIMEOUT specifically in the test, but this
wasn't being honored due to the reasons in the previous commit.

* chore: remove .env file and add go test timeout back